% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/elixr_engine.r
\name{knitr_elixir_engine}
\alias{knitr_elixir_engine}
\title{knitr engine for the Elixir language}
\usage{
knitr_elixir_engine(options)
}
\arguments{
\item{options}{knitr options (required parameter)}
}
\description{
Elixir is a dynamic, functional language designed for building scalable and
maintainable applications.
}
\details{
Elixir leverages the Erlang VM, known for running low-latency, distributed
and fault-tolerant systems, while also being successfully used in web
development and the embedded software domain.
}
\author{
Wendy Smoak
}
\references{
\url{http://elixir-lang.org/}
}

