% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gnuplot_engine.R
\name{knitr_gnuplot_engine}
\alias{knitr_gnuplot_engine}
\title{knitr engine for the Go language}
\usage{
knitr_gnuplot_engine(options)
}
\arguments{
\item{options}{knitr options (required parameter)}
}
\description{
Gnuplot is a portable command-line driven graphing utility for Linux, OS/2,
MS Windows, OSX, VMS, and many other platforms
}
\author{
Bob Rudis
}
\references{
\url{http://www.gnuplot.info/}
}

