% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/go_engine.r
\name{knitr_go_engine}
\alias{knitr_go_engine}
\title{knitr engine for the Go language}
\usage{
knitr_go_engine(options)
}
\arguments{
\item{options}{knitr options (required parameter)}
}
\description{
Go is an open source programming language that makes it easy to build simple,
reliable, and efficient software
}
\author{
Bob Rudis
}
\references{
\url{https://golang.org/}
}

