% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pygments-engine.r
\name{knitr_pygments_engine}
\alias{knitr_pygments_engine}
\title{knitr engine for the pygments syntax highlighting (w/o execution)}
\usage{
knitr_pygments_engine(options)
}
\arguments{
\item{options}{knitr options (required parameter)}
}
\description{
Pygments is a generic syntax highlighter suitable for use in code hosting,
forums, wikis or other applications that need to prettify source code. This
knitr engine only highlights code, it will not execute it.
}
\details{
This requires setting of two options, \code{pyg.ext} which should be a file
extension recognized by the pygments lexer (i.e. \code{py} for Python) and
an optional styling (\code{pyg.sty}) for the code output. One of \code{autumn},
\code{borland}, \code{bw}, \code{colorful}, \code{default}, \code{emacs},
\code{friendly}, \code{fruity}, \code{github}, \code{manni}, \code{monokai},
\code{murphy}, \code{native}, \code{pastie}, \code{perldoc}, \code{tango},
\code{trac}, \code{vim}, \code{vs}, \code{zenburn}.

It will use \code{default} if none are
specified. Styles nabbed from \url{https://github.com/richleland/pygments-css}.

You can install \code{pygments} via \code{pip install pygments}.

There is currently a limitation of one highlight style per knitr document.
}
\note{
\code{echo} should be \code{FALSE} and \code{results} should be \code{asis}.
      However this function will automagically add them if they aren't there. Also,
      this function requrires the system program \code{pygmentize} to be on the
      executable \code{PATH}. It won't work without it.
}
\author{
Bob Rudis
}
\references{
\url{http://pygments.org/}
}

